add patchlist component
